EntityManager: Fix issue where triggers fired when they shouldn't have.
Systems used to require all registered components to exist on an entity before the notify_created trigger fired, but this was broken a while ago. This commit un-breaks that.
notify_destroyed was renamed to notify_will_destroy, which makes it possible to use the entity which is about to be destroyed in the callback.